coproc/tests: Move all copro fixtures to coproc/tests/fixtures
