+++ src/v/coproc/tests/failure_recovery_tests.cc
@@ -8,9 +8,9 @@
  * https://github.com/vectorizedio/redpanda/blob/master/licenses/rcl.md
  */
 
-#include "coproc/tests/utils/coproc_test_fixture.h"
+#include "coproc/tests/fixtures/coproc_test_fixture.h"
+#include "coproc/tests/fixtures/router_test_fixture.h"
 #include "coproc/tests/utils/coprocessor.h"
-#include "coproc/tests/utils/router_test_fixture.h"
 #include "model/fundamental.h"
 #include "model/namespace.h"
 #include "storage/tests/utils/random_batch.h"
