+++ src/v/coproc/tests/fixtures/coproc_test_fixture.h
